Change vault exec shell to sh.
In the RH realeased container image, which is ubi based the default shell is
bash. However in the non ubi based,this will be sh.
Without this, the ansible tasks will fail with
"bash": executable file not found in $PATH: on the non RH (ubi) vault
container image.
Defaulting the shell commands to sh will ensure a bit better compatibility in
different base build vault images, and would support a wider range of
vault images.